+++ dev/tools/controllerbuilder/template/controller.go
@@ -305,3 +305,21 @@ func (a *Adapter) Delete(ctx context.Context) (bool, error) {
 	}
 	return true, nil
 }
+
+func SetStatus(u *unstructured.Unstructured, typedStatus any) error {
+	status, err := runtime.DefaultUnstructuredConverter.ToUnstructured(typedStatus)
+	if err != nil {
+		return fmt.Errorf("error converting status to unstructured: %w", err)
+	}
+
+	old, _, _ := unstructured.NestedMap(u.Object, "status")
+	if old != nil {
+		status["conditions"] = old["conditions"]
+		status["observedGeneration"] = old["observedGeneration"]
+		status["externalRef"] = old["externalRef"]
+	}
+
+	u.Object["status"] = status
+
+	return nil
+}
