+++ dev/tools/controllerbuilder/template/externalresource.go
@@ -24,7 +24,7 @@ import (
 
 const (
 	// TODO(user): Add service domain
-	serviceDomain = "//{{.KindToLower}}.googleapis.com"
+	serviceDomain = "//{{.Service}}.googleapis.com"
 )
 
 // TODO(user): Define resource identity
@@ -53,9 +53,30 @@ func (c *{{.Kind}}Identity) ExternalRef() *string {
 // BuildIDFromExternal builds a {{.Kind}}Identity from a external reference
 func BuildIDFromExternal(external string) (*{{.Kind}}Identity, error) {
 	// TODO(user): Build resource identity from external reference
+	if !strings.HasPrefix(externalRef, serviceDomain) {
+		return nil, fmt.Errorf("externalRef should have prefix %s, got %s", serviceDomain, externalRef)
+	}
+	path := strings.TrimPrefix(externalRef, serviceDomain+"/")
+	tokens := strings.Split(path, "/")
+
+	// TODO(user): Confirm the format of your resources, and verify it like the example below
+	if len(tokens) != 6 || tokens[0] != "projects" || tokens[2] != "locations" || tokens[4] != "{{.KindToLower}}s" {
+		return nil, fmt.Errorf("externalRef should be %s/projects/<project>/locations/<location>/{{.KindToLower}}s/<{{.KindToLower}}>, got %s",
+			serviceDomain, externalRef)
+	}
+	return &{{.Kind}}Identity{
+		project:    tokens[1],
+		location:   tokens[3],
+		{{.KindToLower}}: tokens[5],
+	}, nil
 }
 
-// BuildID builds a {{.Kind}}Identity from resource components
+// BuildID builds a unique identifier {{.Kind}}Identity from resource components
 func BuildID(project, location string) *{{.Kind}}Identity {
 	// TODO(user): Build resource identity from resource components, i.e. project, location, resource id
+	return &{{.Kind}}Identity{
+		project:    project,
+		location:   location,
+		{{.KindToLower}}: {{.KindToLower}},
+	}
 }
