+++ pkg/controller/direct/cloudbuild/workerpool_controller.go
@@ -207,7 +207,7 @@ func (a *Adapter) Create(ctx context.Context, u *unstructured.Unstructured) erro
 		return mapCtx.Err()
 	}
 	status.ExternalRef = a.id.ExternalRef()
-	return direct.SetStatus(u, status)
+	return setStatus(u, status)
 }
 
 func (a *Adapter) Update(ctx context.Context, u *unstructured.Unstructured) error {
@@ -291,7 +291,7 @@ func (a *Adapter) Update(ctx context.Context, u *unstructured.Unstructured) erro
 	if mapCtx.Err() != nil {
 		return fmt.Errorf("update workerpool status %w", mapCtx.Err())
 	}
-	return direct.SetStatus(u, status)
+	return setStatus(u, status)
 }
 
 func (a *Adapter) Export(ctx context.Context) (*unstructured.Unstructured, error) {
@@ -317,4 +317,22 @@ func (a *Adapter) Delete(ctx context.Context) (bool, error) {
 		}
 	}
 	return true, nil
+}
+
+func setStatus(u *unstructured.Unstructured, typedStatus any) error {
+	status, err := runtime.DefaultUnstructuredConverter.ToUnstructured(typedStatus)
+	if err != nil {
+		return fmt.Errorf("error converting status to unstructured: %w", err)
+	}
+
+	old, _, _ := unstructured.NestedMap(u.Object, "status")
+	if old != nil {
+		status["conditions"] = old["conditions"]
+		status["observedGeneration"] = old["observedGeneration"]
+		status["externalRef"] = old["externalRef"]
+	}
+
+	u.Object["status"] = status
+
+	return nil
 }
